Multiple timestamps modifications
Added Timestamps table in the database. Added comments in the
DatabaseHelper class wherever it needs changes to support multiple
timestamps. I kept the timestamp column in the Words table in order to
keep it backward compatible (and gain some time in order to not quering
for the last timestamp of a word).